+++ app/src/main/java/eu/kanade/tachiyomi/data/notification/NotificationReceiver.kt
@@ -618,6 +618,25 @@ class NotificationReceiver : BroadcastReceiver() {
         }
 
         /**
+         * Returns [PendingIntent] that opens the failed updates screen.
+         *
+         * @param context context of application
+         * @return [PendingIntent]
+         */
+        internal fun openFailedUpdatesPendingActivity(context: Context): PendingIntent {
+            val intent = Intent(context, MainActivity::class.java).apply {
+                action = Constants.SHORTCUT_FAILED_UPDATES
+                flags = Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_SINGLE_TOP
+            }
+            return PendingIntent.getActivity(
+                context,
+                0,
+                intent,
+                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE,
+            )
+        }
+
+        /**
          * Returns [PendingIntent] that cancels a backup restore job.
          *
          * @param context context of application