+++ app/src/main/java/eu/kanade/tachiyomi/ui/updates/UpdatesScreenModel.kt
@@ -44,6 +44,7 @@ import tachiyomi.domain.library.service.LibraryPreferences
 import tachiyomi.domain.manga.interactor.GetManga
 import tachiyomi.domain.manga.model.applyFilter
 import tachiyomi.domain.source.service.SourceManager
+import tachiyomi.domain.updates.interactor.GetMangaUpdateErrors
 import tachiyomi.domain.updates.interactor.GetUpdates
 import tachiyomi.domain.updates.model.UpdatesWithRelations
 import tachiyomi.domain.updates.service.UpdatesPreferences
@@ -61,6 +62,7 @@ class UpdatesScreenModel(
     private val getManga: GetManga = Injekt.get(),
     private val getChapter: GetChapter = Injekt.get(),
     private val libraryPreferences: LibraryPreferences = Injekt.get(),
+    private val getMangaUpdateErrors: GetMangaUpdateErrors = Injekt.get(),
     private val updatesPreferences: UpdatesPreferences = Injekt.get(),
     val snackbarHostState: SnackbarHostState = SnackbarHostState(),
 ) : StateScreenModel<UpdatesScreenModel.State>(State()) {
@@ -119,6 +121,13 @@ class UpdatesScreenModel(
                 .collect(this@UpdatesScreenModel::updateDownloadState)
         }
 
+        screenModelScope.launchIO {
+            getMangaUpdateErrors.subscribeCount()
+                .catch { logcat(LogPriority.ERROR, it) }
+                .collectLatest { count ->
+                    mutableState.update { it.copy(failedUpdatesCount = count) }
+                }
+        }
         getUpdatesItemPreferenceFlow()
             .map { prefs ->
                 listOf(
@@ -450,6 +459,7 @@ class UpdatesScreenModel(
         val hasActiveFilters: Boolean = false,
         val items: List<UpdatesItem> = listOf(),
         val dialog: Dialog? = null,
+        val failedUpdatesCount: Long = 0,
     ) {
         val selected = items.filter { it.selected }
         val selectionMode = selected.isNotEmpty()